Extract ExcelLinter test and introduce pytest.
