+++ pyproject.toml
@@ -16,6 +16,7 @@ pandas = "^1.3.3"
 chardet = "^4.0.0"
 jeraconv = "^0.2.1"
 openpyxl = "^3.0.9"
+pytest = "^6.2.5"
 
 [tool.poetry.dev-dependencies]
 yapf = "^0.31.0"
